Version 1.1
Now supports changing of pixel size like in the original
OlcPixelGameEngine.

To make use of the pixel size changing one has to use the drawing
functions from RweGameEngine and not the normal drawing functions in
Graphics2D.

Anti aliasing is on by default when using drawing functions in
Graphics2D and off by default when using the drawing functions in
RweGameEngine. It can also be turned on for these functions though with
"useAntiAliasing(boolean toggle)".